requirements.txt
pyrogram>=2.0.106
python-dotenv>=1.1.1
tgcrypto>=1.2.5